requirements.txt
jsonschema==4.15.0
PyYAML==6.0
requests==2.28.1
